+++ src/main/java/com/insane96mcp/shieldsplus/setup/SPEnchantments.java
@@ -1,10 +1,7 @@
 package com.insane96mcp.shieldsplus.setup;
 
 import com.insane96mcp.shieldsplus.ShieldsPlus;
-import com.insane96mcp.shieldsplus.world.item.enchantment.ShieldAblazeEnchantment;
-import com.insane96mcp.shieldsplus.world.item.enchantment.ShieldRecoilEnchantment;
-import com.insane96mcp.shieldsplus.world.item.enchantment.ShieldReflectionEnchantment;
-import com.insane96mcp.shieldsplus.world.item.enchantment.ShieldReinforcedEnchantment;
+import com.insane96mcp.shieldsplus.world.item.enchantment.*;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
@@ -17,4 +14,7 @@ public class SPEnchantments {
     public static final RegistryObject<Enchantment> REFLECTION = ENCHANTMENTS.register("reflection", ShieldReflectionEnchantment::new);
     public static final RegistryObject<Enchantment> REINFORCED = ENCHANTMENTS.register("reinforced", ShieldReinforcedEnchantment::new);
     public static final RegistryObject<Enchantment> ABLAZE = ENCHANTMENTS.register("ablaze", ShieldAblazeEnchantment::new);
+    public static final RegistryObject<Enchantment> LIGHTWEIGHT = ENCHANTMENTS.register("lightweight", ShieldLightweightEnchantment::new);
+    public static final RegistryObject<Enchantment> FAST_RECOVERY = ENCHANTMENTS.register("fast_recovery", ShieldFastRecoveryEnchantment::new);
+    public static final RegistryObject<Enchantment> SHIELD_BASH = ENCHANTMENTS.register("shield_bash", ShieldBashEnchantment::new);
 }