+++ src/main/java/com/insane96mcp/shieldsplus/world/item/enchantment/ShieldFastRecoveryEnchantment.java
@@ -1,0 +1,33 @@
+package com.insane96mcp.shieldsplus.world.item.enchantment;
+
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.ShieldItem;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+
+public class ShieldFastRecoveryEnchantment extends Enchantment {
+
+	public static final int TICKS = 12;
+
+	public ShieldFastRecoveryEnchantment() {
+		super(Rarity.RARE, EnchantmentCategory.BREAKABLE, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
+	}
+
+	public int getMinCost(int p_44598_) {
+		return 22;
+	}
+
+	public int getMaxCost(int p_44600_) {
+		return this.getMinCost(p_44600_) + 22;
+	}
+
+	public int getMaxLevel() {
+		return 1;
+	}
+
+	@Override
+	public boolean canApplyAtEnchantingTable(ItemStack stack) {
+		return stack.getItem() instanceof ShieldItem;
+	}
+}