+++ changelog.md
@@ -1,11 +1,13 @@
 # Changelog
 
 ## Upcoming
-* Added 3 shields enchantments
+* Added 6 shields enchantments
   * Reinforced: increases damage blocked by 0.5 per level (max IV)
   * Recoil: Knockbacks entities and launches back projectiles that hit the shield (max II)
   * Reflection: Damages entities that hit the shield by 15% per level of the damage taken (max III)
   * Ablaze: Sets the entities that hit the shield on fire for 4 seconds per level (max II)
+  * Lightweight: Makes the player move faster when blocking
+  * Fast Recovery: Makes the Shield cooldown 0.6 seconds less
   * Shield Bash
 * Shields can now be enchanted at the Enchanting Table